// SmartGrid/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    primaryLookupName: ComponentFramework.PropertyTypes.StringProperty;
    primaryEntitySet: ComponentFramework.PropertyTypes.StringProperty;
    smartGridDataSet: ComponentFramework.PropertyTypes.DataSet;
}
export interface IOutputs {
}